+++ src/joi/index.ts
@@ -1,4 +1,4 @@
 export { resolveJSONSchema, resolveBundledJSONSchema } from './resolve';
-export { ResolveOptions as Options, SubSchemas } from './options';
+export { ResolveOptions, FormatOptions } from './options';
 export { generateJoiStatement, JoiStatement } from './generate';
 export { formatJoi } from './format';